requirements.txt
flask==2.0.1
pdf2image==1.16.0
numpy==1.24.3
opencv-python-headless==4.5.4.60
torch==1.13.1
torchvision==0.14.1
easyocr==1.6.2
pandas==2.1.3
werkzeug==2.0.3
gunicorn==20.1.0
python-dotenv==1.0.0
